refactor: 할 일 체크에 대한 커스텀훅 분리 #212
